refactor: Indicate env in load test file names
