+++ k6/index.js
@@ -1,4 +1,4 @@
 // @ts-nocheck - TS complains about duplicate options exports. This is necessary
 // to have one entry point to generate all JS files.
-export * from "./graphql/components";
-export * from "./graphql/observations";
+export * from "./graphql/test-components";
+export * from "./graphql/test-observations";
